Official release v11.3.0

+++ Projects/NUCLEO-F401RE/Applications/IKS01A3/VibrationMonitoring/Src/main.c
@@ -775,7 +775,7 @@ static uint8_t Meas_Odr(void)
     AcceleroODR.Period = 1 / AcceleroODR.Frequency;
   }
 
-  AcceleroODR.Tau = exp(-(float)(1000 * AcceleroODR.Period) / MotionSP_Parameters.tau);
+  AcceleroODR.Tau = (float)exp(-(double)(1000.0f * AcceleroODR.Period) / MotionSP_Parameters.tau);
 
   /* Disable accelerometer to avoid interrupt conflicts on highest ODRs */
   if (IKS01A3_MOTION_SENSOR_Disable(IKS01A3_LSM6DSO_0, MOTION_ACCELERO) != BSP_ERROR_NONE)
